+++ parser/src/databases/txout_index_to_amount.rs
@@ -1,15 +1,15 @@
 use std::{
     collections::BTreeMap,
-    mem,
+    fs, mem,
     ops::{Deref, DerefMut},
 };
 
 use allocative::Allocative;
-use rayon::prelude::*;
+use itertools::Itertools;
 
 use crate::structs::{Amount, Date, Height, TxoutIndex};
 
-use super::{AnyDatabaseGroup, Database as _Database, Metadata};
+use super::{AnyDatabase, AnyDatabaseGroup, Database as _Database, Metadata};
 
 type Key = TxoutIndex;
 type Value = Amount;
@@ -95,21 +95,57 @@ impl AnyDatabaseGroup for TxoutIndexToAmount {
         }
     }
 
-    fn export(&mut self, height: Height, date: Date) -> color_eyre::Result<()> {
-        mem::take(&mut self.map)
-            .into_par_iter()
-            .try_for_each(|(_, db)| db.export())?;
-
-        self.metadata.export(height, date)?;
-
-        Ok(())
-    }
-
     fn reset_metadata(&mut self) {
         self.metadata.reset();
     }
 
     fn folder<'a>() -> &'a str {
         "txout_index_to_amount"
+    }
+
+    fn open_all(&mut self) {
+        let path = Self::full_path();
+
+        let folder = fs::read_dir(path);
+
+        if folder.is_err() {
+            return;
+        }
+
+        folder
+            .unwrap()
+            .map(|entry| {
+                entry
+                    .unwrap()
+                    .path()
+                    .file_name()
+                    .unwrap()
+                    .to_str()
+                    .unwrap()
+                    .to_owned()
+            })
+            .filter(|file_name| file_name.contains(".."))
+            .for_each(|path| {
+                self.open_db(
+                    &path
+                        .split("..")
+                        .next()
+                        .unwrap()
+                        .parse::<u64>()
+                        .unwrap()
+                        .into(),
+                );
+            });
+    }
+
+    fn drain_to_vec(&mut self) -> Vec<Box<dyn AnyDatabase + Send>> {
+        mem::take(&mut self.map)
+            .into_values()
+            .map(|db| Box::new(db) as Box<dyn AnyDatabase + Send>)
+            .collect_vec()
+    }
+
+    fn export_metadata(&mut self, height: Height, date: Date) -> color_eyre::Result<()> {
+        self.metadata.export(height, date)
     }
 }