+++ parser/src/structs/txout_index.rs
@@ -9,6 +9,9 @@ pub struct TxoutIndex {
 }
 direct_repr!(TxoutIndex);
 
+const SHIFT: u64 = 16;
+const AND: u64 = (1 << SHIFT) - 1;
+
 impl TxoutIndex {
     #[inline(always)]
     pub fn new(tx_index: u32, vout: u16) -> Self {
@@ -17,6 +20,15 @@ impl TxoutIndex {
 
     #[inline(always)]
     pub fn as_u64(&self) -> u64 {
-        ((self.tx_index as u64) << 16_u64) + self.vout as u64
+        ((self.tx_index as u64) << SHIFT) + self.vout as u64
+    }
+}
+
+impl From<u64> for TxoutIndex {
+    fn from(value: u64) -> Self {
+        Self {
+            tx_index: (value >> SHIFT) as u32,
+            vout: (value & AND) as u16,
+        }
     }
 }